+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-18
+- `21:46` -- **Updated** `apps/desktopProbe/src/server/overlayBrowserView.ts`
 - `21:45` -- **Updated** `apps/desktopProbe/src/server/overlayBrowserView.ts`
 - `21:44` -- **Updated** `apps/desktopProbe/src/server/overlayBrowserView.ts`
 - `21:44` -- **Added** `apps/desktopProbe/src/server/logger.ts`